Addons: Move extract-source to storybook/addons public API
