+++ addons/docs/src/blocks/enhanceSource.ts
@@ -1,6 +1,5 @@
 import { combineParameters } from '@storybook/client-api';
-import { StoryContext, Parameters } from '@storybook/addons';
-import { extractSource, LocationsMap } from '@storybook/source-loader';
+import { StoryContext, Parameters, extractSource, LocationsMap } from '@storybook/addons';
 
 interface StorySource {
   source: string;
