+++ lib/addons/src/public_api.ts
@@ -10,5 +10,6 @@ export * from './index';
 export * from './types';
 export * from './storybook-channel-mock';
 export * from './hooks';
+export * from './source';
 
 export default addons;
