+++ lib/addons/src/source.ts
@@ -1,4 +1,18 @@
-import type { SourceBlock } from './types';
+export interface SourceLoc {
+  line: number;
+  col: number;
+}
+
+export interface SourceBlock {
+  startBody: SourceLoc;
+  endBody: SourceLoc;
+  startLoc: SourceLoc;
+  endLoc: SourceLoc;
+}
+
+export interface LocationsMap {
+  [key: string]: SourceBlock;
+}
 
 /**
  * given a location, extract the text from the full source
