+++ lib/source-loader/src/abstract-syntax-tree/generate-helpers.js
@@ -1,4 +1,5 @@
 import { storyNameFromExport, sanitize } from '@storybook/csf';
+import { extractSource } from '@storybook/addons';
 import mapKeys from 'lodash/mapKeys';
 import { patchNode } from './parse-helpers';
 import getParser from './parsers';
@@ -9,7 +10,6 @@ import {
   popParametersObjectFromDefaultExport,
   findExportsMap as generateExportsMap,
 } from './traverse-helpers';
-import { extractSource } from '../extract-source';
 
 export function sanitizeSource(source) {
   return JSON.stringify(source)
