+++ lib/source-loader/src/index.ts
@@ -4,5 +4,3 @@ import { transform } from './build';
 export default transform;
 
 export * from './types';
-
-export { extractSource } from './extract-source';
